Add MetaTagsPart under MetaTags feature
Introduced a new feature for managing meta tags on a content item named
"MetaTags". Enabling this feature will add a `MetaTagsPart` that can
be attached to a content item. The plan is for the `Data` property on
the `MetaTagsPart` to contain the meta content as JSON. When the content
item is displayed, the JSON will be parsed and rendered in the the
output via the `IResourceManager`.

Currently the editor experience is a textarea where the user can input
the JSON. The plan is to add an easy to use UI to manage the meta
content.